loans search by unique id

+++ loan/nestjs-back/src/loans/application/usecase/loans.delete.use.case.ts
@@ -1,28 +1,24 @@
-import { Inject, Injectable } from "@nestjs/common";
+import { BadRequestException, Inject, Injectable } from "@nestjs/common";
 import { LoansDeleteAdaptorOutputDto } from "../../outbound/dtos/adaptor/loans.delete.adaptor.output.dto";
 import { LoansDeleteAdaptor } from "../../domain/adaptor/loans.delete.adaptor";
 import { LoansDeleteAdaptorInputDto } from "../../inbound/dtos/adaptor/loans.delete.adaptor.input.dto";
-import { LoansValidateRequiredLoanUniqueIdInterface } from "../../domain/interface/loans.validate.required.loan.unique.id.interface";
-import { LoansExistsLoanUniqueIdInterface } from "../../domain/interface/loans.exists.loan.unique.id.interface";
-import { LoansValidateRequiredLoanCreditorUniqueIdInterface } from "../../domain/interface/loans.validate.required.loan.creditor.unique.id.interface";
 import { LoansExistsLoanCreditorUniqueIdInterface } from "../../domain/interface/loans.exists.loan.creditor.unique.id.interface";
-import { LoansValidateRequiredLoanDebtorUniqueIdInterface } from "../../domain/interface/loans.validate.required.loan.debtor.unique.id.interface";
 import { LoansExistsLoanDebtorUniqueIdInterface } from "../../domain/interface/loans.exists.loan.debtor.unique.id.interface";
+import {
+  CREDITOR_UNIQUE_ID_REQUIRED,
+  DEBTOR_UNIQUE_ID_REQUIRED,
+  UNIQUE_ID_REQUIRED,
+} from "../../../_common/constants/http/errors/400";
+import { LoansExistsLoanUniqueIdInterface } from "../../domain/interface/loans.exists.loan.unique.id.interface";
 
 @Injectable()
 export class LoansDeleteUseCase implements LoansDeleteAdaptor {
   constructor(
     @Inject("DELETE") private readonly repository: LoansDeleteAdaptor,
-    @Inject("VALIDATE_REQUIRED_LOAN_UNIQUE_ID")
-    private readonly compareDbUniqueIdWith: LoansValidateRequiredLoanUniqueIdInterface,
     @Inject("EXISTS_LOAN_UNIQUE_ID")
     private readonly compareExistsDbUniqueIdWith: LoansExistsLoanUniqueIdInterface,
-    @Inject("VALIDATE_REQUIRED_LOAN_CREDITOR_UNIQUE_ID")
-    private readonly compareDbCreditorUniqueIdWith: LoansValidateRequiredLoanCreditorUniqueIdInterface,
     @Inject("EXISTS_LOAN_CREDITOR_UNIQUE_ID")
     private readonly compareExistsDbCreditorUniqueIdWith: LoansExistsLoanCreditorUniqueIdInterface,
-    @Inject("VALIDATE_REQUIRED_LOAN_DEBTOR_UNIQUE_ID")
-    private readonly compareDbDebtorUniqueIdWith: LoansValidateRequiredLoanDebtorUniqueIdInterface,
     @Inject("EXISTS_LOAN_DEBTOR_UNIQUE_ID")
     private readonly compareExistsDbDebtorUniqueIdWith: LoansExistsLoanDebtorUniqueIdInterface
   ) {}
@@ -32,22 +28,27 @@ export class LoansDeleteUseCase implements LoansDeleteAdaptor {
   ): Promise<LoansDeleteAdaptorOutputDto> {
     const { id, debtorUniqueId, creditorUniqueId } = dto;
 
-    await this.compareDbUniqueIdWith.validateRequiredLoanUniqueId({ id });
-    await this.compareExistsDbUniqueIdWith.existsLoanUniqueId({ id });
+    const {
+      response: { existsLoanUniqueId },
+    } = await this.compareExistsDbUniqueIdWith.existsLoanUniqueId({ id });
+    if (existsLoanUniqueId) throw new BadRequestException(UNIQUE_ID_REQUIRED);
 
-    await this.compareDbDebtorUniqueIdWith.validateRequiredLoanDebtorUniqueId({
+    const {
+      response: { existsLoanDebtorUniqueId },
+    } = await this.compareExistsDbDebtorUniqueIdWith.existsLoanDebtorUniqueId({
       debtorUniqueId,
     });
-    await this.compareExistsDbDebtorUniqueIdWith.existsLoanDebtorUniqueId({
-      debtorUniqueId,
-    });
+    if (existsLoanDebtorUniqueId)
+      throw new BadRequestException(DEBTOR_UNIQUE_ID_REQUIRED);
 
-    await this.compareDbCreditorUniqueIdWith.validateRequiredLoanCreditorUniqueId(
-      { creditorUniqueId }
-    );
-    await this.compareExistsDbCreditorUniqueIdWith.existsLoanCreditorUniqueId({
-      creditorUniqueId,
-    });
+    const {
+      response: { existsLoanCreditorUniqueId },
+    } =
+      await this.compareExistsDbCreditorUniqueIdWith.existsLoanCreditorUniqueId(
+        { creditorUniqueId }
+      );
+    if (existsLoanCreditorUniqueId)
+      throw new BadRequestException(CREDITOR_UNIQUE_ID_REQUIRED);
 
     return await this.repository.delete(dto);
   }
